change double fields to decimal

+++ DTOs/ProductFilterDTO.cs
@@ -7,8 +7,8 @@ namespace OnlineShop.DTOs
 {
     public class ProductFilterDTO
     {
-        public float? MinPrice {get;set;}
-        public float? MaxPrice { get; set; }
+        public decimal? MinPrice {get;set;}
+        public decimal? MaxPrice { get; set; }
         //public float? MinRating {get;set;}
         //public float? MaxRating {get;set;}
         public int? CategoryId  {get;set;}
